varningar a sant shit lolbert

+++ Source/Memory/TestScenarios.h
@@ -117,7 +117,7 @@ void VerifyMTStack(unsigned int _numAllocations, unsigned int _dataTypeSize, uns
 					if (memoryAddress == verifyMTStack_MemoryPointers[i][j])
 					{
 						printf("Memory has been found in more than one place (%iz). Locations: (Thread: %i, Memory Index: %i) and (Thread: %i, Memory Index: %i)\n", 
-							memoryAddress,
+							(int)memoryAddress,
 							threadIndex, 
 							memoryIndex,
 							i,
